drivers/requirements.txt
numpy
Cython
PyTango>=8.1.1
pytest>=2.6.4
pyserial>=2.6
